Refactor delegate-account and link-device routes
Use component views for consistency with the other routes
