+++ src/lib/views.ts
@@ -1,3 +1,7 @@
 export type BackupView = 'backup' | 'are-you-sure'
 
 export type ConnectView = 'connect' | 'open-connected-device'
+
+export type DelegateAccountView = 'connect-backup-device' | 'delegate-account'
+
+export type LinkDeviceView = 'link-device' | 'load-filesystem'
